# openapiDoc.JSON
{
  "openapi": "3.0.3",
  "info": {
    "title": "Groupomania social network",
    "description": "This is the documentation of Groupomania social network API.",
    "termsOfService": "http://groupomania.com/terms/",
    "contact": {
      "name": "API support",
      "url": "http://groupomania.com/contact/",
      "email": "contactAPI@groupomania.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [{
    "url": "http://localhost:{port}/{basePath}",
    "variables": {
      "port": {
        "default": "3001"
      },
      "basePath": {
        "default": "api"
      }
    }
  }],
...
  "paths": {
...
    "/reactions": {
      "post": {
        "tags": ["reactions"],
        "summary": "Create a reaction.",
        "description": "User creates a reaction.",
        "operationId": "createOneReaction",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Reaction properties. Exactly one of 'postId' or 'commentId' must be completed.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reactionPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "reaction created."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "get": {
        "tags": ["reactions"],
        "summary": "Get all reactions.",
        "description": "User gets all reactions.",
        "operationId": "getAllReactions",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "minDate",
            "in": "query",
            "description": "Get all reactions who have been created after this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "maxDate",
            "in": "query",
            "description": "Get all reactions who have been created before this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Get a limited amount of reactions.",
            "schema": {
              "type": "number",
              "example": 2
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Returns elements from oldest to newest ('asc') or from newest to oldest ('desc').",
            "schema": {
              "type": "string",
              "example": "desc"
            }
          },
          {
            "name": "fromUserId",
            "in": "query",
            "description": "Get all reactions from a specific user.",
            "schema": {
              "type": "string",
              "example": "632dc722c75414645a459eb2"
            }
          },
          {
            "name": "fromPostId",
            "in": "query",
            "description": "Get all reactions from a specific post.",
            "schema": {
              "type": "string",
              "example": "632efd854148fdc69cd7e652"
            }
          },
          {
            "name": "fromCommentId",
            "in": "query",
            "description": "Get all reactions from a specific comment.",
            "schema": {
              "type": "string",
              "example": "632efd838148fdc69cd7e652"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/reaction"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["reactions"],
        "summary": "Delete all reactions.",
        "description": "Admin deletes all reactions.",
        "operationId": "deleteAllReactions",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "400": {
            "description": "Bad request."
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/reactions/{reactionId}": {
      "get": {
        "tags": ["reactions"],
        "summary": "Get one reaction.",
        "description": "User gets properties of a reaction.",
        "operationId": "getOneReaction",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "reactionId",
            "in": "path",
            "description": "Id of reaction to return.",
            "schema": {
              "type": "string",
              "example": "632efdbb4148fdc69cd7e784"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/reaction"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "put": {
        "tags": ["reactions"],
        "summary": "Modify one reaction.",
        "description": "User modifies a reaction of his own, or admin modifies a reaction.",
        "operationId": "modifyOneReaction",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "reactionId",
            "in": "path",
            "description": "Id of reaction to modify.",
            "schema": {
              "type": "string",
              "example": "632efdbb4148fdc69cd7e784"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "New properties of the reaction.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reactionPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/reaction"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["reactions"],
        "summary": "Delete one reaction.",
        "description": "User deletes his own reaction, or admin deletes a reaction.",
        "operationId": "deleteOneReaction",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "reactionId",
            "in": "path",
            "description": "Id of reaction to delete.",
            "schema": {
              "type": "string",
              "example": "632efdbb4148fdc69cd7e784"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "reaction": {
        "title": "Reaction",
        "description": "Reaction schema. Exactly one of 'postId' or 'commentId' must be completed.",
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "632efdbb4148fdc69cd7e784"
          },
          "type": {
            "type": "string",
            "example": "Like"
          },
          "postId": {
            "type": "string",
            "example": ""
          },
          "commentId": {
            "type": "string",
            "example": "632efdbb4148fdc69cd7e654"
          },
          "userId": {
            "type": "string",
            "example": "632dc722c75414645a459eb2"
          },
          "creationDate": {
            "type": "string",
            "example": "2022-13-01T14:48:02.809Z"
          },
          "__v": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "_id",
          "type",
          "userId",
          "creationDate",
          "__v"
        ]
      },
...
      "bearerToken": {
        "description": "Bearer token schema.",
        "type": "string",
        "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzMxZDk1MWU3N2YxNDAxMmQxMjExNmEiLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNjY0MjExMzEyLCJleHAiOjE2NjQyOTc3MTJ9.03h-xFcgHmwpYCHMZq36hnSAsHNS7TItboYmAliKSIk"
      },
...
      "reactionPost": {
        "title": "ReactionPost",
        "description": "Reaction schema used when user publishes a reaction. Exactly one of 'postId' or 'commentId' must be completed.",
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Like"
          },
          "postId": {
            "type": "string",
            "example": "632efd854148fdc69cd7e652"
          },
          "commentId": {
            "type": "string",
            "example": "632efd854148fdc69cd7e652"
          }
        },
        "example": {
          "type" : "like",
          "postId" : "632efd854148fdc69cd7e652"
        },
        "required": [
          "type"
        ]
      },
      "reactionPut": {
        "title": "ReactionPut",
        "description": "Reaction schema used when user modifies his reaction.",
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Dislike"
          }
        },
        "required": [
          "type"
        ]
      },
...
    }
  }
}